+++ ThirdEyeVIDemoTests/Properties/AssemblyInfo.cs
