+++ ThirdEyeVIDemoTests/ThirdEyeVIDemoTests.csproj
@@ -36,7 +36,12 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="Emgu.CV, Version=2.4.10.1939, Culture=neutral, PublicKeyToken=7281126722ab4438, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>.\Emgu.CV.dll</HintPath>
+    </Reference>
     <Reference Include="System" />
+    <Reference Include="System.Drawing" />
   </ItemGroup>
   <Choose>
     <When Condition="('$(VisualStudioVersion)' == '10.0' or '$(VisualStudioVersion)' == '') and '$(TargetFrameworkVersion)' == 'v3.5'">
@@ -51,14 +56,51 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
+    <Compile Include="FrameProcessorTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="WebcamFetcherTests.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\WristbandCsharp\ThirdEyeVIDemo.csproj">
+    <ProjectReference Include="..\CAPIServerTest\CAPIServerTest.csproj">
+      <Project>{7dc40e51-3a0a-4d97-adfa-663e1b2ec98d}</Project>
+      <Name>CAPIServerTest</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\CAPIStreamClient\CAPIStreamClient.csproj">
+      <Project>{e2c96291-8aab-4089-9f2a-03311107048b}</Project>
+      <Name>CAPIStreamClient</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\CAPIStreamCommon\CAPIStreamCommon.csproj">
+      <Project>{79d59ff1-dadf-4eed-8321-c36c38d65646}</Project>
+      <Name>CAPIStreamCommon</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\CAPIStreamServer\CAPIStreamServer.csproj">
+      <Project>{15818dac-2d3c-4d64-a7c0-0f88499ce793}</Project>
+      <Name>CAPIStreamServer</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\CAPIStreamTest\CAPIStreamTest.csproj">
+      <Project>{4da31846-37cd-4e6b-ae28-a3fb9eba68f7}</Project>
+      <Name>CAPIStreamTest</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\CMT_Tracker\Clustering\Clustering.csproj">
+      <Project>{f5c7f496-de23-4ecd-916c-5be8b4bdff33}</Project>
+      <Name>Clustering</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\CMT_Tracker\CMT_Tracker\CMT_Tracker.csproj">
+      <Project>{041772d0-10c9-4f02-86d0-e2e985f48d38}</Project>
+      <Name>CMT_Tracker</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\ThirdEyeVIDemo\ThirdEyeVIDemo.csproj">
       <Project>{CD5327FB-3B05-4D84-8A3E-C95A3228F3F7}</Project>
       <Name>ThirdEyeVIDemo</Name>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="EmguDLLs\Emgu.CV.dll" />
+    <Content Include="EmguDLLs\Emgu.CV.UI.dll" />
+    <Content Include="EmguDLLs\Emgu.CV.UI.xml" />
+    <Content Include="EmguDLLs\Emgu.CV.xml" />
+    <Content Include="EmguDLLs\Emgu.Util.dll" />
+    <Content Include="EmguDLLs\Emgu.Util.xml" />
   </ItemGroup>
   <Choose>
     <When Condition="'$(VisualStudioVersion)' == '10.0' And '$(IsCodedUITest)' == 'True'">
